Update the interface
