+++ examples/phi-3.5-mini-to-coreml/test_script.py
