+++ src/latched/model_exporters/auto.py
@@ -14,5 +14,5 @@ class AutoExporter(BaseModelExporter):
     """
 
     @classmethod
-    def run(cls, model_wrapper: BaseModelWrapper, output_name: str = "model.abc") -> None:
+    def run(cls, model_wrapper: BaseModelWrapper, **kwargs) -> None:
         raise NotImplementedError
