+++ src/latched/model_exporters/openvino.py
@@ -15,10 +15,11 @@ class OpenVINOExporter(BaseModelExporter):
     """
 
     @classmethod
-    def run(cls, model_wrapper: BaseModelWrapper, output_name: str = "openvino_model") -> None:
+    def run(cls, model_wrapper: BaseModelWrapper, **kwargs) -> None:
         if isinstance(model_wrapper, HuggingFaceModelWrapper):
             from optimum.exporters.openvino import export_from_model
 
+            output_name = kwargs.get("output_name", "ov_model")
             export_from_model(model_wrapper.original_model, output_name)
 
             print(f"Model exported to {output_name}")
