+++ src/latched/models/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2024 TBD Labs Inc. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
